restructure descriptors into folders
file size was getting too large. Sub-modules for descriptors makes sense
with intention to add more; video, hub, otg, dfu etc.
